==> notebooks/1.0-data-collecting.ipynb <==
<div style="text-align: center; background-color: #559cff; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Final Project - Programming For Data Science @ FIT-HCMUS, VNU-HCM 📌
</div>

<div style="text-align: center; background-color: #b1d1ff; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Stage 01 - Data Collecting
</div>

The data we choose are movies and their features on the IMDb website (**https://www.imdb.com/**). Choosing a data analysis topic such as a movie on the IMDb website can bring many benefits and research opportunities. Link to get the dataset: **https://www.kaggle.com/datasets/elvinrustam/imdb-movies-dataset**
- **Rich Data Collection Ability**: 
IMDb is one of the major sources of information about movies, with thousands of movies and detailed information about actors, directors, genres, revenues, and many other factors. Data from IMDb can provide a rich and diverse dataset for analysis.
- **Market and Competitive Analysis**:
Data from IMDb provides the opportunity to analyze the film market and competition between films. You can evaluate the performance of films, identify factors that determine success, and predict future trends.
- **Understanding Audience Interests and Trends**:
We can analyze audience preferences through their reviews, views, and opinions. A better understanding of popular trends can help filmmakers and film producers understand audience needs.
