==> notebooks/1.0-preprocessing.ipynb <==
<div style="text-align: center; background-color: #559cff; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Lab 02 - Introduction To Data Science @ FIT-HCMUS, VNU-HCM 📌
</div>

<div style="text-align: center; background-color: #b1d1ff; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Stage 02 - Preprocessing
</div>

## Import

import os
import sys

module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import datetime
import re

## Exploring Data

### Read raw data from file

raw_df = pd.read_csv('../data/raw/IMDbMovies.csv')

# TEST
raw_df.head()

### How many rows and how many columns does the raw data have?

Next, I calculate the number of rows and columns of the DataFrame `raw_df` and store it in the variable `shape` (tuple). **If my data has the number of rows smaller than 1000, please crawl the data once again with larger year interval**.

shape = raw_df.shape 

# TEST
print(f"Current shape: {shape}")

if shape[0] > 1000:
    print(f"Your data good!.")
else:
    print(f"Your raw data absolutely small. Please choose larger year interval.!")

### What does each row mean?


Each row indicates the values ​​corresponding to eachcolumn.n

### Does the raw data have duplicate rows?

Next, I calculate the number of rows with duplicate indexes and store it in the variable `num_duplicated_rows`. In a group of lines with the same index, the first line is not counted as repeated.

duplicated_df = raw_df[raw_df.duplicated()]
num_duplicated_rows = duplicated_df.shape[0]

# TEST
if num_duplicated_rows == 0:
    print(f"Raw data has no duplicated line.!")
else:
    if num_duplicated_rows > 1:
        ext = "lines"
    else:
        ext = "line"
    print(f"Raw data has {num_duplicated_rows} duplicated " + ext + ". Please de-deduplicate raw data.!")

# De-deduplicate raw data
raw_df = raw_df.drop_duplicates()
raw_df.head()

### What does each column mean?

I don't have the meaning of each column available, so I will parse it from `kaggle`. 

import requests as req
from bs4 import BeautifulSoup

features = raw_df.columns.values.tolist()
descriptions = []

url = 'https://www.kaggle.com/datasets/elvinrustam/imdb-movies-dataset'
response = req.get(url)
soup = BeautifulSoup(response.text, 'html.parser')
needed_text = str(soup.find_all('script', type = "application/ld+json")[0])
needed_text = needed_text.split('\\n\\n')

for feature in features:
    for item in needed_text:
        item = item.replace('amp;', '')
        if f'**{feature}:**' in item:
            values = item.replace('**', '').split(':')

            desc = values[1].strip()
            if feature == 'Motion Picture Rating':
                idx = needed_text.index(item)
                categories = needed_text[(idx + 2) : (idx + 7)]
                
                for cate in categories:
                    cate = cate.strip().replace('**', '')
                    desc += '\n' + cate
                    
            descriptions.append(desc)
            break
meaning_col = pd.DataFrame(list(dict(zip(features, descriptions)).items()), columns = ['Feature', 'Description'])
meaning_col

### What data type does each column currently have? Are there any columns whose data types are not suitable for further processing?

In this step, I calculate the data type (dtype) of each column in DataFrame `raw_df` and save the result into Series `dtypes` (This Series has the index as the column name).

dtypes = raw_df.dtypes

# TEST
dtypes

We observe that some columns have unsuitable data types or special presentations such as `Runtime`, `Release Year`, `Rating`, etc. Therefore, we should convert them and normalize some columns with special presentations. Columns have to convert:
- `Runtime` : numeric (minutes)
- `Release Year` : string (because `Year` is a categorical attribute).
- `Rating`, `Number of Ratings`, `Budget`, `Gross in US & Canada`, `Gross worldwide`, `Opening Weekend Gross in US & Canada` : numeric

#### Convert `Runtime` column

def convertRuntime(inp):
    if inp is np.nan:
        return inp
    hour = 0
    min = 0
    for item in inp.split(' '):
        if 'h' in item:
            hour = int(item[:-1])
        elif 'm' in item:
            min = int(item[:-1])
    return hour * 60 + min
    
raw_df['Runtime'] = raw_df['Runtime'].apply(convertRuntime)
raw_df.rename(columns = {'Runtime' : 'Runtime (Minutes)'}, inplace = True)
raw_df['Runtime (Minutes)']

#### Convert `Release Year` column

raw_df['Release Year'] = raw_df['Release Year'].astype('str')
raw_df['Release Year']

#### Convert `Ratings` and `Number of Ratings` columns

raw_df['Rating'] = raw_df['Rating'].apply(lambda x: None if x is np.nan else float(x[:-3]))
raw_df.rename(columns = {'Rating' : 'Rating (Out of 10)'}, inplace = True)
raw_df['Rating (Out of 10)']

raw_df['Number of Ratings'] = raw_df['Number of Ratings'].apply(lambda x: float(x.replace('K', ''))
                                                                if pd.notna(x) and 'K' in x else
                                                                float(x.replace('M', '')) * 1000
                                                                if pd.notna(x) and 'M' in x else 
                                                                pd.to_numeric(x, errors = 'coerce'))
raw_df.rename(columns = {'Number of Ratings' : 'Number of Ratings (in thousands)'}, inplace = True)
raw_df['Number of Ratings (in thousands)']

#### Convert Columns of Currency Unit

Notes that column `Opening Weekend Gross in US & Canada` has both `Currency` and `Datetime` in each value. Therefore, we should remove the `Datetime` part because it's not important, and keep the `Currency` sequence to convert.

def convertCurrency(inp):
    if inp is np.nan:
        return None
    inp = inp.replace(',', '')
    return pd.to_numeric(re.search(r'\d+', inp).group()) / 1000000
    
raw_df['Budget (in milions)'] = raw_df['Budget'].apply(convertCurrency)
raw_df['Gross in US & Canada (in milions)'] = raw_df['Gross in US & Canada'].apply(convertCurrency)
raw_df['Gross worldwide (in milions)'] = raw_df['Gross worldwide'].apply(convertCurrency)
raw_df['Opening Weekend Gross in US & Canada (in milions)'] = raw_df['Opening Weekend Gross in US & Canada'].apply(convertCurrency)
raw_df = raw_df.drop(columns = ['Budget', 'Gross in US & Canada', 'Gross worldwide', 'Opening Weekend Gross in US & Canada'], axis = 1)

# Result after converting
raw_df.dtypes

# TEST
dtypes = raw_df.dtypes
float_cols = set(dtypes[(dtypes==np.float32) | (dtypes==np.float64)].index)
assert len(float_cols) == 7
object_cols = set(dtypes[dtypes == object].index)
assert len(object_cols) == 7

### For each column with numeric data type, how are the values distributed?

For columns with numeric data types, I will calculate:
- Percentage of missing values
- The min
- The lower quartile
- The median
- The upper quartile
- The max

I will save the results to a DataFrame `num_col_info_df`, where:
- The names of the columns are the names of the numeric columns in `raw_df`
- Names of rows: "missing_ratio", "min", "lower_quartile", "median", "upper_quartile", "max"  

For ease of viewing, I round all values to 1 decimal place using the `.round(1)` method.

def missing_ratio(df : pd.Series):
    return (df.isnull().sum() * 100 / len(raw_df)).round(1)
def lower_quartile(df : pd.Series):
    return (df.quantile(q = 0.25)).round(1)
def median(df : pd.Series):
    return (df.quantile(q = 0.5)).round(1)
def upper_quartile(df : pd.Series):
    return (df.quantile(q = 0.75)).round(1)

num_col_info_df = raw_df.select_dtypes(exclude = 'object')
num_col_info_df = num_col_info_df.agg([missing_ratio, "min", lower_quartile, median, upper_quartile, "max"])
num_col_info_df

# Visualize (OPTIONAL)

dict(num_col_info_df.iloc[0])

After identifying the basic statistical numbers that describe my data, I further need to determine the features that have a large number of missing values. Such features are not useful for the analysis stage and must be removed from the dataset.

Depending on my goals, the threshold for "large" can be defined. Usually, if the percentage of missing values is greater than 50%, the column is dropped from the dataframe and an updated dataframe is returned.

def drop_missing_features(df: pd.DataFrame, missing_lst: dict = dict(num_col_info_df.iloc[0]), threshold: float = 50.0) -> pd.DataFrame:
    for col in df.columns:
        if df[col].dtype != np.float64 or df[col].dtype != np.float32:
            continue
        elif missing_lst[col] > threshold:
            df = df.drop(col, axis = 1)
    return df

raw_df = drop_missing_features(raw_df)

# TEST
raw_df.head()

After remove features which have large missing values, our dataframe still have missing values. So that, we need to fill these missing values so that they can be used in analysis.

def filling_missing_value(df: pd.DataFrame) -> pd.DataFrame:
    for col in df.columns:
        for value in df[col]:
            if value is None:
                value = np.NaN
    for col in df.select_dtypes(include = 'number').columns:
        mean = df[col].mean().round(1)
        df[col] = df[col].fillna(mean)
    return df

raw_df = filling_missing_value(raw_df)

# TEST
raw_df.head()

# TEST
raw_df.select_dtypes(exclude='object').agg([missing_ratio, "min", lower_quartile, median, upper_quartile, "max"])

### For each column with a non-numeric data type, how are the values distributed?

For columns with non-numeric data types, I calculate:
- Percentage (from 0 to 100) of missing values
- Number of values (the values here are different values and we do not consider missing values): with columns whose type is categorical, it is a set with a finite number of categories. Directly counting the number of values in these columns doesn't make much sense, so it's better to count the number of elements of all types.
- The percentage (from 0 to 100) of each value is sorted by decreasing percentage (we do not consider missing values, the ratio is the ratio compared to the number of non-missing values): you use a dictionary to store , key is the value, value is the percentage; With the column corresponding to each type, the method is similar to above.

I will save the results to DataFrame `cat_col_info_df`, where:
- The names of the columns are the names of the non-numeric columns in `raw_df`
- The names of the lines are: "missing_ratio", "num_values", "value_ratios"

For ease of viewing, I round all values to 3 decimal place using the `.round(3)` method.

pd.set_option('display.max_colwidth', 100) # For clearly
pd.set_option('display.max_columns', None) # For clearly

cat_col_info_df = raw_df.select_dtypes(include = 'object')

def num_values(ser : pd.Series):
    return len(ser.unique())
def value_ratios(ser : pd.Series) -> dict:
    return (ser.value_counts(dropna = True) * 100 / len(ser)).round(3).to_dict()

cat_col_info_df = cat_col_info_df.agg([missing_ratio, num_values, value_ratios])
cat_col_info_df

# Visualize (OPTIONAL)

### Is the collected data reasonable?

**Question 1**: Is there any value in the `Rating (Out of 10)` column that exceeds 10?

def is_out_of_10(ser: pd.Series):
    results = ser > 10
    return results.value_counts()[False]

# TEST
assert is_out_of_10(raw_df['Rating (Out of 10)']) == len(raw_df)

The test is True, so there is no value in `Rating (Out of 10)` column that exceeds 10. This column is reasonable.

**Question 2**: Ares`Gross in US & Canada (in milions)`'s values greateress th`Opening Weekend Gross in US & Canada (in milions)`'s valuesher?

def checkout_gross(ser1: pd.Series, ser2: pd.Series):
    result = ser1 >= ser2
    return result.value_counts()[False]

# TEST
assert checkout_gross(raw_df['Gross in US & Canada (in milions)'], raw_df['Opening Weekend Gross in US & Canada (in milions)']) > 0

### Save processed data

print(f"Total number of features: {raw_df.shape[1]}")
raw_df.dtypes

raw_df.to_csv(f"../data/processed/IMDbMovies_processed.csv", index = False)
