==> notebooks/4.0-data-modelling.ipynb <==
<div style="text-align: center; background-color: #559cff; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Final Project - Programming For Data Science @ FIT-HCMUS, VNU-HCM 📌
</div>

<div style="text-align: center; background-color: #b1d1ff; font-family: 'Trebuchet MS', Arial, sans-serif; color: white; padding: 20px; font-size: 40px; font-weight: bold; border-radius: 0 0 0 0; box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2);">
  Stage 4.0 - Data modelling
</div>

**In this part, we will use regression model Random Forest to predict movie ratings based on Main Genres, Motion Picture Rating, Runtime, Release Year, Number of Ratings, Budget, Gross in US & Canada, Gross worldwide, Opening Weekend Gross in US & Canada.**

## Import

import warnings
warnings.filterwarnings('ignore')

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

import sklearn
from sklearn.preprocessing import LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, r2_score

## Read data

pd.set_option('display.max_columns', None)
movie_df = pd.read_csv('./../data/processed/IMDbMovies_processed.csv')
movie_df.head()

## Feature engineering

df = movie_df[['Main Genres', 'Motion Picture Rating', 'Runtime (Minutes)', 
               'Release Year', 'Rating (Out of 10)', 'Number of Ratings (in thousands)',
               'Budget (in milions)', 'Gross in US & Canada (in milions)', 'Gross worldwide (in milions)',
               'Opening Weekend Gross in US & Canada (in milions)']].dropna()

#Transform Main Genre and Motion Picture Rating using one-hot encoding and label encoding
one_hot_df = pd.DataFrame()
def splitGenre(s):
    return s.split(',')
genre_stack = df['Main Genres'].apply(splitGenre).agg(pd.Series).stack().agg(pd.Series).stack()
genre_dummy = pd.get_dummies(genre_stack).groupby(level = 0).sum()
genre_dummy.columns = ['Genre_' + col for col in genre_dummy.columns]
df = pd.concat([genre_dummy, df], axis = 1)

label_encoder=LabelEncoder()
df['Motion Picture Rating (encoded)'] = label_encoder.fit_transform(df['Motion Picture Rating'])

df = df.drop(['Main Genres', 'Motion Picture Rating'], axis=1)
df

## Data preparation

X = df.drop('Rating (Out of 10)', axis=1)  # Features
y = df['Rating (Out of 10)']  # Target

# Split the data into training and test sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

## Training model & Prediction

model = RandomForestRegressor()
model.fit(X_train, y_train)
y_pred = model.predict(X_test)

## Evaluation
- We use 2 metrics to evaluate the efficiency of model performances:
    - Mean squared error: estimates the average of the squares of the errors.
    - R-squared - Coefficient of determination: the proportion of change in the dependent variable that can be predicted from the independent variable.

mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)
print("Mean squared error (MSE): ", mse)
print("R-squared: ", r2)

## Results analysis

Analysis results: *mse* is 0.4 and *r-squared* is 0.5, these are acceptable results for the model, showing that the model works quite well in predicting ratings based on many other factors.
